telebot/requirements.txt
APScheduler>=3.6.3
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///tmp/build/80754af9/beautifulsoup4_1631874778482/work
Bottleneck @ file:///C:/ci/bottleneck_1648029530934/work
brotlipy>=0.7.0
certifi>=2020.6.20
cffi @ file:///C:/ci_310/cffi_1643036627027/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cryptography @ file:///C:/ci/cryptography_1643209086516/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
cytoolz>=0.11.0
dask>=2.14.0
debugpy @ file:///C:/ci_310/debugpy_1642079916595/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
entrypoints>=0.3
et-xmlfile>=1.1.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fonttools>=4.25.0
html5lib @ file:///Users/ktietz/demo/mc3/conda-bld/html5lib_1629144453894/work
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
imagecodecs @ file:///C:/ci_310/imagecodecs_1642082516950/work
imageio>=2.8.0
imagesize>=1.2.0
install>=1.3.4
ipykernel @ file:///C:/ci/ipykernel_1647000977835/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1646948607209/work
jdcal @ file:///Users/ktietz/demo/mc3/conda-bld/jdcal_1630584345063/work
jedi @ file:///C:/ci/jedi_1644315428305/work
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1643638337975/work
jupyter-core @ file:///C:/ci/jupyter_core_1646994639510/work
kiwisolver @ file:///C:/ci_310/kiwisolver_1642083321497/work
lxml>=4.5.0
matplotlib @ file:///C:/ci/matplotlib-suite_1647423638658/work
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mkl-fft>=1.3.1
mkl-random @ file:///C:/ci_310/mkl_random_1643050563308/work
mkl-service>=2.4.0
munkres>=1.1.4
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
networkx>=2.4
numexpr @ file:///C:/ci_310/numexpr_1642084889105/work
numpy @ file:///C:/ci_310/numpy_and_numpy_base_1643798589088/work
openpyxl>=3.0.3
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///C:/ci/pandas_1647955074650/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow>=9.0.1
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
pytesseract>=0.3.3
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-telegram-bot>=13.0
pytz>=2021.3
PyWavelets @ file:///C:/ci/pywavelets_1648728096239/work
pywin32>=302
pyzmq @ file:///C:/ci_310/pyzmq_1642090360481/work
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
scikit-image @ file:///C:/ci/scikit-image_1648214343198/work
scipy @ file:///C:/ci_310/scipy_1642247684282/work
selenium>=3.141.0
sip>=4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve @ file:///tmp/build/80754af9/soupsieve_1636706018808/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tifffile @ file:///tmp/build/80754af9/tifffile_1627275862826/work
toolz @ file:///tmp/build/80754af9/toolz_1636545406491/work
tornado @ file:///C:/ci_310/tornado_1642093111997/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
tzlocal>=2.1
unicodecsv>=0.14.1
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings>=0.5.1
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
wincertstore>=0.2
